refactored packaging
