+++ app/src/main/java/c4stor/com/feheroes/model/hero/MovementType.java
@@ -1,6 +1,6 @@
 package c4stor.com.feheroes.model.hero;
 
-import c4stor.com.feheroes.model.skill.Inheritance;
+import c4stor.com.feheroes.model.Inheritance;
 
 /**
  * Created by eclogia on 14/06/17.
