+++ app/src/main/java/c4stor/com/feheroes/model/skill/WeaponColor.java
@@ -1,5 +1,7 @@
 package c4stor.com.feheroes.model.skill;
 
+import c4stor.com.feheroes.model.Inheritance;
+
 /**
  * Created by eclogia on 13/06/17.
  */
